+++ Data/ExtendedRoadUpgrades.cs
@@ -1,18 +1,21 @@
-﻿using ExtendedRoadUpgrades.Models;
-using Game.Prefabs;
-using System.Collections.Generic;
-
-namespace ExtendedRoadUpgrades.Data
+﻿namespace ExtendedRoadUpgrades.Data
 {
+    using System.Collections.Generic;
+    using Game.Prefabs;
+    using global::ExtendedRoadUpgrades.Models;
+
     /// <summary>
     /// Main container of all the available upgrade modes.
     /// This is not intended as an example of how to design a proper C# project so I'll
     /// just dump everything I need into a static variable and call it a day.
-    /// 
-    /// Please don't use my code to learn how to program! :)
+    ///
+    /// Please don't use my code to learn how to program! :).
     /// </summary>
     internal class ExtendedRoadUpgrades
     {
+        /// <summary>
+        ///     This variable contains all the available upgrade modes that we support.
+        /// </summary>
         public static IEnumerable<ExtendedRoadUpgradeModel> Modes = new[]
         {
             // Quay
@@ -23,36 +26,36 @@ namespace ExtendedRoadUpgrades.Data
                     {
                         "en-US",
                         "Quay"
-                    }
+                    },
                 },
                 Description = new Dictionary<string, string> {
                     {
                         "en-US",
                         "A quay, if you installed this mod you know what it is :)"
-                    }
+                    },
                 },
                 m_SetUpgradeFlags = new CompositionFlags
                 {
-                    m_Right = CompositionFlags.Side.Raised
+                    m_Right = CompositionFlags.Side.Raised,
                 },
                 m_UnsetUpgradeFlags = new CompositionFlags
                 {
                     m_General = CompositionFlags.General.Elevated,
-                    m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Lowered
+                    m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Lowered,
                 },
 
                 // TODO: not sure how this works yet
                 m_SetState = new[]
                 {
-                    NetPieceRequirements.Raised
+                    NetPieceRequirements.Raised,
                 },
                 m_UnsetState = new[]
                 {
                     NetPieceRequirements.Lowered,
                     NetPieceRequirements.Elevated,
                     NetPieceRequirements.LowTransition,
-                    NetPieceRequirements.OppositeLowTransition
-                }
+                    NetPieceRequirements.OppositeLowTransition,
+                },
             },
 
             // Retaining Wall
@@ -63,36 +66,36 @@ namespace ExtendedRoadUpgrades.Data
                     {
                         "en-US",
                         "Retaining Wall"
-                    }
+                    },
                 },
                 Description = new Dictionary<string, string> {
                     {
                         "en-US",
                         "A retaining wall, if you installed this mod you know what it is :)"
-                    }
+                    },
                 },
                 m_SetUpgradeFlags = new CompositionFlags
                 {
-                    m_Right = CompositionFlags.Side.Lowered
+                    m_Right = CompositionFlags.Side.Lowered,
                 },
                 m_UnsetUpgradeFlags = new CompositionFlags
                 {
                     m_General = CompositionFlags.General.Elevated,
-                    m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised
+                    m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised,
                 },
 
                 // TODO: not sure how this works yet
                 m_SetState = new[]
                 {
-                    NetPieceRequirements.Lowered
+                    NetPieceRequirements.Lowered,
                 },
                 m_UnsetState = new[]
                 {
                     NetPieceRequirements.Raised,
                     NetPieceRequirements.Elevated,
                     NetPieceRequirements.LowTransition,
-                    NetPieceRequirements.OppositeLowTransition
-                }
+                    NetPieceRequirements.OppositeLowTransition,
+                },
             },
 
             // Elevated
@@ -103,37 +106,77 @@ namespace ExtendedRoadUpgrades.Data
                     {
                         "en-US",
                         "Elevated"
-                    }
+                    },
                 },
                 Description = new Dictionary<string, string> {
                     {
                         "en-US",
                         "Elevated mode, kind of similar to bridges"
-                    }
+                    },
                 },
                 m_SetUpgradeFlags = new CompositionFlags
                 {
-                    m_General = CompositionFlags.General.Elevated
+                    m_General = CompositionFlags.General.Elevated,
                 },
                 m_UnsetUpgradeFlags = new CompositionFlags
                 {
                     m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised | CompositionFlags.Side.Lowered,
-                    m_Left = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised | CompositionFlags.Side.Lowered
+                    m_Left = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised | CompositionFlags.Side.Lowered,
                 },
 
                 // TODO: not sure how this works yet
                 m_SetState = new[]
                 {
-                    NetPieceRequirements.Elevated
+                    NetPieceRequirements.Elevated,
                 },
                 m_UnsetState = new[]
                 {
                     NetPieceRequirements.Raised,
                     NetPieceRequirements.Lowered,
                     NetPieceRequirements.LowTransition,
-                    NetPieceRequirements.OppositeLowTransition
-                }
-            }
+                    NetPieceRequirements.OppositeLowTransition,
+                },
+            },
+
+            // Tunnel
+            new ExtendedRoadUpgradeModel
+            {
+                Id = "Tunnel",
+                Name = new Dictionary<string, string> {
+                    {
+                        "en-US",
+                        "Tunnel"
+                    },
+                },
+                Description = new Dictionary<string, string> {
+                    {
+                        "en-US",
+                        "Tunnel mode, it might not look perfect but it works."
+                    },
+                },
+                m_SetUpgradeFlags = new CompositionFlags
+                {
+                    m_General = CompositionFlags.General.Tunnel,
+                },
+                m_UnsetUpgradeFlags = new CompositionFlags
+                {
+                    m_Right = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised | CompositionFlags.Side.Lowered,
+                    m_Left = CompositionFlags.Side.WideSidewalk | CompositionFlags.Side.PrimaryBeautification | CompositionFlags.Side.Raised | CompositionFlags.Side.Lowered,
+                },
+
+                // TODO: not sure how this works yet
+                m_SetState = new[]
+                {
+                    NetPieceRequirements.Tunnel,
+                },
+                m_UnsetState = new[]
+                {
+                    NetPieceRequirements.Raised,
+                    NetPieceRequirements.Lowered,
+                    NetPieceRequirements.LowTransition,
+                    NetPieceRequirements.OppositeLowTransition,
+                },
+            },
         };
     }
 }