+++ Patches/SystemOrder_Initialize.cs
@@ -1,11 +1,14 @@
-﻿
-using Game.Common;
-using Game;
-using HarmonyLib;
-using ExtendedRoadUpgrades.Systems;
-
-namespace ExtendedRoadUpgrades.Patches
+﻿namespace ExtendedRoadUpgrades.Patches
 {
+    using ExtendedRoadUpgrades.Systems;
+    using Game;
+    using Game.Common;
+    using HarmonyLib;
+
+    /// <summary>
+    ///     This patch hooks into the initial <see cref="SystemOrder.Initialize(Game.UpdateSystem)"/> to attach out custom
+    ///     <see cref="GameSystemBase"/> to the Update workflow.
+    /// </summary>
     [HarmonyPatch(typeof(SystemOrder), "Initialize")]
     class SystemOrder_Initialize
     {
@@ -17,8 +20,8 @@ namespace ExtendedRoadUpgrades.Patches
             updateSystem.UpdateAt<PickerToolSystem>(SystemUpdatePhase.ToolUpdate);
 #endif
             Plugin.Logger.LogInfo($"[{nameof(SystemOrder_Initialize)}.{nameof(Postfix)}] Creating retaining walls fixer system...");
-            updateSystem.World.GetOrCreateSystem<NodeRetainingWallUpdateSystem>();
-            updateSystem.UpdateAt<NodeRetainingWallUpdateSystem>(SystemUpdatePhase.ToolUpdate);
+            updateSystem.World.GetOrCreateSystem<NodeUpdateFixerSystem>();
+            updateSystem.UpdateAt<NodeUpdateFixerSystem>(SystemUpdatePhase.ToolUpdate);
 
             Plugin.Logger.LogInfo($"...done.");
         }